Moved EditorPage component into its own module
